% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{fastener}
\alias{fastener}
\title{A function to create the object of Type:FasteneR}
\usage{
fastener(db_name = db_name, project_name = project_name, plot_list)
}
\arguments{
\item{str}{name of database and the name of the particular project"}
}
\value{
A vector contain all valid tokens in the original input string
}
\description{
A function to create the object of Type:FasteneR
}
