% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{fr_add_product}
\alias{fr_add_product}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
fr_add_product(.data, description)
}
\arguments{
\item{str}{location"}
}
\value{
A vector contain all valid tokens in the original input string
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}
