% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_developer_scripts.R
\name{fr_fork_clone}
\alias{fr_fork_clone}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
fr_fork_clone()
}
\arguments{
\item{str}{location"}
}
\value{
repo
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}
