% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{fr_get_resulting_dataframe}
\alias{fr_get_resulting_dataframe}
\title{A Function to disconnect to connection to database and returns resulting dataframe}
\usage{
fr_get_resulting_dataframe(.data)
}
\arguments{
\item{str}{date"}
}
\value{
resulting dataframe
}
\description{
A Function to disconnect to connection to database and returns resulting dataframe
}
