% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{fr_get_transaction}
\alias{fr_get_transaction}
\title{A Function to clean a single input string by removing punctuation and numbers and tokenizing it.}
\usage{
fr_get_transaction(.data, date)
}
\arguments{
\item{str}{date"}
}
\value{
fastener object
}
\description{
A Function to clean a single input string by removing punctuation and numbers and tokenizing it.
}
